chore: rename package
example -> chess
