+++ app/src/main/java/org/chess/App.java
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example;
+package org.chess;
 
 public class App {
     public String getGreeting() {
