+++ app/src/test/java/org/chess/AppTest.java
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example;
+package org.chess;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
